Added the internal formula constructor.
