+++ crates/syntax/Cargo.toml
@@ -17,8 +17,9 @@ workspace = true
 
 [dependencies]
 merc_lts.workspace = true
-merc_refinement.workspace = true
 merc_pest_consume.workspace = true
+merc_reduction.workspace = true
+merc_refinement.workspace = true
 merc_utilities.workspace = true
 
 itertools.workspace = true
